Fix non-Apple build; hoist nested types; unify build scripts
- Replace OutputCache's concrete ProcessInfo dep with an EnvironmentProvider
  protocol so FakeEnvironment (struct) replaces the FakeProcessInfo subclass
  that broke on swift-corelibs-foundation (final class, environment on an
  extension). Clears the OutputCacheTests failure on Linux/Windows/Android.
- Add swift-system as a conditional dep + canImport(System)/SystemPackage
  shim so the skit target compiles on Linux/Windows.
- Hoist Runner.SetupError -> RunnerSetupError, Runner.ToolchainVerification
  -> ToolchainVerification, Skit.Run.CommandError -> RunCommandError,
  DirectoryRender.FileOutcome -> FileOutcome.
- Replace DirectoryRender with [FileOutcome] + a failureCount extension.
- Lift render/renderSingle/renderBatch out of extension Skit.Run into
  enum Render (none used self).
- Move SKIT_LIB_DIR lookup to ProcessInfo.skitLibPath.
- Inject FileManager via @Sendable () -> FileManager on Runner; thread
  through Runner+Session and Bundle.resolveLibPath.
- Delete Docs/skit-internals.md + Docs/string-literal-audit.md.
- Unify Scripts/build-skit-{debug,release}.sh into Scripts/build-skit.sh
  (release default; --debug flag).
- Lint cleanups: split FileManager+FileSystem.swift into PathKind /
  FileFingerprint / FileEnumerationError; extract RenderTaskResult to
  its own file; reorder type contents (cases/properties/init/methods);
  fix conditional-returns-on-newline + multiline-arguments-brackets;
  add doc comment to ContentHasher.update(data:). Lint goes from 21
  violations to 5 (the remaining 5 are all pre-existing).
