+++ Sources/SyntaxKit/Execution/EnvironmentProvider.swift
@@ -1,5 +1,5 @@
 //
-//  RunError.swift
+//  EnvironmentProvider.swift
 //  SyntaxKit
 //
 //  Created by Leo Dion.
@@ -27,22 +27,18 @@
 //  OTHER DEALINGS IN THE SOFTWARE.
 //
 
-/// Typed error surfaced by `Runner`. It decouples the renderer from any
-/// particular caller: `Runner` reports *what* went wrong, and the caller
-/// (CLI, build plugin, in-process driver) decides how to present it.
-public enum RunError: Error {
-  /// The input path was invalid — missing, or a directory given without an
-  /// output directory.
-  case invalidInput(String)
-  /// Single-file render: the spawned `swift` exited non-zero. Carries that
-  /// code (e.g. a compile failure, `124` on timeout, `128 + signal`) and
-  /// the (path-rewritten) stderr the toolchain emitted, so the caller can
-  /// surface diagnostics without having to fish them out elsewhere. Also
-  /// carries the session's `toolchainVerification`: when it isn't `.verified`,
-  /// the failure may stem from a Swift-toolchain mismatch the check couldn't
-  /// rule out, which the caller can hint at.
-  case renderFailed(exitCode: Int32, stderr: String, toolchain: ToolchainVerification)
-  /// A wrapped Foundation/Subprocess failure (file read/write, spawn error)
-  /// that has no dedicated mapping.
-  case unexpected(any Error)
+public import Foundation
+
+/// The slice of `ProcessInfo` `OutputCache` depends on. Abstracted into a
+/// protocol so tests can inject a fixed environment without subclassing
+/// `ProcessInfo` — `swift-corelibs-foundation` declares `ProcessInfo` as
+/// `final` with `environment` on an extension, so subclass-and-override
+/// doesn't compile on Linux/Windows.
+public protocol EnvironmentProvider: Sendable {
+  /// Process environment variables.
+  var environment: [String: String] { get }
+  /// Identifier of the current process, used to namespace staging dirs.
+  var processIdentifier: Int32 { get }
 }
+
+extension ProcessInfo: EnvironmentProvider {}
