+++ Sources/SyntaxKit/Execution/FileFingerprint.swift
@@ -1,5 +1,5 @@
 //
-//  RunError.swift
+//  FileFingerprint.swift
 //  SyntaxKit
 //
 //  Created by Leo Dion.
@@ -27,22 +27,13 @@
 //  OTHER DEALINGS IN THE SOFTWARE.
 //
 
-/// Typed error surfaced by `Runner`. It decouples the renderer from any
-/// particular caller: `Runner` reports *what* went wrong, and the caller
-/// (CLI, build plugin, in-process driver) decides how to present it.
-public enum RunError: Error {
-  /// The input path was invalid — missing, or a directory given without an
-  /// output directory.
-  case invalidInput(String)
-  /// Single-file render: the spawned `swift` exited non-zero. Carries that
-  /// code (e.g. a compile failure, `124` on timeout, `128 + signal`) and
-  /// the (path-rewritten) stderr the toolchain emitted, so the caller can
-  /// surface diagnostics without having to fish them out elsewhere. Also
-  /// carries the session's `toolchainVerification`: when it isn't `.verified`,
-  /// the failure may stem from a Swift-toolchain mismatch the check couldn't
-  /// rule out, which the caller can hint at.
-  case renderFailed(exitCode: Int32, stderr: String, toolchain: ToolchainVerification)
-  /// A wrapped Foundation/Subprocess failure (file read/write, spawn error)
-  /// that has no dedicated mapping.
-  case unexpected(any Error)
+import Foundation
+
+/// Size + modification date of a file: the change-detection inputs behind a
+/// content fingerprint/stamp.
+internal struct FileFingerprint: Equatable {
+  /// File size in bytes (0 when the attribute is absent).
+  internal let size: Int
+  /// Last-modification date (Unix epoch when the attribute is absent).
+  internal let modificationDate: Date
 }
