+++ Sources/SyntaxKit/Execution/ToolchainVerification.swift
@@ -1,5 +1,5 @@
 //
-//  RunError.swift
+//  ToolchainVerification.swift
 //  SyntaxKit
 //
 //  Created by Leo Dion.
@@ -27,22 +27,19 @@
 //  OTHER DEALINGS IN THE SOFTWARE.
 //
 
-/// Typed error surfaced by `Runner`. It decouples the renderer from any
-/// particular caller: `Runner` reports *what* went wrong, and the caller
-/// (CLI, build plugin, in-process driver) decides how to present it.
-public enum RunError: Error {
-  /// The input path was invalid — missing, or a directory given without an
-  /// output directory.
-  case invalidInput(String)
-  /// Single-file render: the spawned `swift` exited non-zero. Carries that
-  /// code (e.g. a compile failure, `124` on timeout, `128 + signal`) and
-  /// the (path-rewritten) stderr the toolchain emitted, so the caller can
-  /// surface diagnostics without having to fish them out elsewhere. Also
-  /// carries the session's `toolchainVerification`: when it isn't `.verified`,
-  /// the failure may stem from a Swift-toolchain mismatch the check couldn't
-  /// rule out, which the caller can hint at.
-  case renderFailed(exitCode: Int32, stderr: String, toolchain: ToolchainVerification)
-  /// A wrapped Foundation/Subprocess failure (file read/write, spawn error)
-  /// that has no dedicated mapping.
-  case unexpected(any Error)
+/// Outcome of the bundle/local Swift-toolchain compatibility check performed
+/// when a render session was created. Stored on the `Runner` and carried on
+/// `RunError.renderFailed` so a caller can hint that an *unverified* toolchain
+/// may be the real cause of a build error — swiftmodules aren't reliably
+/// compatible across compiler versions. A confirmed *mismatch* never reaches
+/// here; it fails session setup via `RunnerSetupError.toolchainMismatch`.
+public enum ToolchainVerification: Sendable, Equatable {
+  /// The bundle's recorded `swift --version` matched the local one.
+  case verified
+  /// The caller opted out of the check (`enforceToolchainCheck == false`).
+  case notChecked
+  /// The check ran but couldn't compare versions — the bundle had no
+  /// toolchain stamp, or the local `swift --version` couldn't be captured —
+  /// so compatibility is unknown.
+  case unverified
 }
