+++ Sources/skit/ProcessInfo+SkitLibPath.swift
@@ -1,5 +1,5 @@
 //
-//  RunError.swift
+//  ProcessInfo+SkitLibPath.swift
 //  SyntaxKit
 //
 //  Created by Leo Dion.
@@ -27,22 +27,15 @@
 //  OTHER DEALINGS IN THE SOFTWARE.
 //
 
-/// Typed error surfaced by `Runner`. It decouples the renderer from any
-/// particular caller: `Runner` reports *what* went wrong, and the caller
-/// (CLI, build plugin, in-process driver) decides how to present it.
-public enum RunError: Error {
-  /// The input path was invalid — missing, or a directory given without an
-  /// output directory.
-  case invalidInput(String)
-  /// Single-file render: the spawned `swift` exited non-zero. Carries that
-  /// code (e.g. a compile failure, `124` on timeout, `128 + signal`) and
-  /// the (path-rewritten) stderr the toolchain emitted, so the caller can
-  /// surface diagnostics without having to fish them out elsewhere. Also
-  /// carries the session's `toolchainVerification`: when it isn't `.verified`,
-  /// the failure may stem from a Swift-toolchain mismatch the check couldn't
-  /// rule out, which the caller can hint at.
-  case renderFailed(exitCode: Int32, stderr: String, toolchain: ToolchainVerification)
-  /// A wrapped Foundation/Subprocess failure (file read/write, spawn error)
-  /// that has no dedicated mapping.
-  case unexpected(any Error)
+import Foundation
+
+extension ProcessInfo {
+  /// Environment variable holding an override for the libSyntaxKit directory.
+  internal static let skitLibDirEnvironmentKey = "SKIT_LIB_DIR"
+
+  /// The libSyntaxKit directory specified via `SKIT_LIB_DIR`, or `nil` when
+  /// the env var is unset or empty.
+  internal var skitLibPath: String? {
+    environment[Self.skitLibDirEnvironmentKey].flatMap { $0.isEmpty ? nil : $0 }
+  }
 }
